"Mario & Luigi: Superstar Saga" volume & panpot events

+++ GBAMusicStudio/MIDI/MIDIKeyboard.cs
@@ -109,7 +109,7 @@ namespace GBAMusicStudio.MIDI
             else if (e.Message.Command == ChannelCommand.NoteOn) // Note on
             {
                 var i = new Instrument();
-                i.Play(new M4ATrack(16) { Voice = vNum, Volume = volume }, note, (byte)(Config.MIDIKeyboardFixedVelocity ? 127 : e.Message.Data2), -1);
+                i.Play(new M4ATrack(16) { Voice = vNum, Volume = (byte)volume }, note, (byte)(Config.MIDIKeyboardFixedVelocity ? 127 : e.Message.Data2), -1);
                 instruments.Add(i);
                 Tick(null, null); // Prevent input delay
             }
